+++ Core/src/main/java/dk/sdu/mmmi/cbse/main/Game.java
@@ -84,14 +84,14 @@ public class Game
 
             sr.begin(ShapeRenderer.ShapeType.Line);
 
-            float[] shapex = entity.getShapeX();
-            float[] shapey = entity.getShapeY();
+            double[] shapex = entity.getShapeX();
+            double[] shapey = entity.getShapeY();
 
             for (int i = 0, j = shapex.length - 1;
                     i < shapex.length;
                     j = i++) {
 
-                sr.line(shapex[i], shapey[i], shapex[j], shapey[j]);
+                sr.line((float)shapex[i], (float)shapey[i], (float)shapex[j], (float)shapey[j]);
             }
 
             sr.end();
